Map errors in the delete pending segments API to proper HTTP status code.
Currently, any errors in the delete pending segments API would result
in the client not knowing about the http status and reason. This patch fixes it.

This is the error from the overlord logs that don't get mapped:
2024-01-12T01:06:41,807 ERROR [qtp1757718624-150] com.sun.jersey.spi.container.ContainerResponse - The RuntimeException could not be mapped to a response, re-throwing to the HTTP container
java.lang.IllegalArgumentException: Cannot delete pendingSegments because there is at least one active task created at 2024-01-12T01:06:41.318Z

Also, clean up the exception messages and add tests.

+++ indexing-service/src/main/java/org/apache/druid/indexing/overlord/IndexerMetadataStorageAdapter.java
@@ -21,6 +21,7 @@ package org.apache.druid.indexing.overlord;
 
 import com.google.common.base.Preconditions;
 import com.google.inject.Inject;
+import org.apache.druid.error.InvalidInput;
 import org.apache.druid.indexer.TaskInfo;
 import org.apache.druid.java.util.common.DateTimes;
 import org.joda.time.DateTime;
@@ -58,11 +59,15 @@ public class IndexerMetadataStorageAdapter
         DateTimes.MAX
     );
 
-    Preconditions.checkArgument(
-        !deleteInterval.overlaps(activeTaskInterval),
-        "Cannot delete pendingSegments because there is at least one active task created at %s",
-        activeTaskInterval.getStart()
-    );
+    if (deleteInterval.overlaps(activeTaskInterval)) {
+      throw InvalidInput.exception(
+        "Cannot delete pendingSegments for datasource[%s] as there's at least one active task with interval[%s] "
+        + "that overlaps with the delete interval[%s]. Please retry when there are no active tasks.",
+        dataSource,
+        deleteInterval,
+        activeTaskInterval
+      );
+    }
 
     return indexerMetadataStorageCoordinator.deletePendingSegmentsCreatedInInterval(dataSource, deleteInterval);
   }
